feat: add nanoid as a dependency and implement ID generator
The commit introduces nanoid package to the project dependencies in the package.json file and in the yarn.lock file. This will be used to generate unique identifiers. A utility function 'generateId' was also added to 'src/common/utils/generate-id.ts'. This function uses nanoid's customAlphabet function to generate 24-character long unique identifiers.