feat: implemented feature interface for GetHelloService
Updated GetHelloService class to implement a newly created interface, GetHelloInterface. This adjustment was made to maintain a high level of code consistency and modularity, also ensuring that MockGetHelloService follows the defined service pattern. Additionally, adjusted the providers in get-hello.module to use the new interface. The same changes were made in get-hello.controller specs to use the mock service.