refactor: move domain interfaces into entities folder
The domain interfaces for various entities including category items, categories, boards, transactions and users have been moved into a new subfolder called entities. This change was made to improve the organization of the code and make it easier to locate specific domain interfaces. In addition to moving the files, the import statements within these files were updated to reflect the new directory structure. This is a part of broader efforts to streamline the project's architecture.