feat: add create board feature
A feature for creating a new board was added to the application. This included the creation of a new CreateBoard module, service and controller in REST presentation, required mappings and repository in infrastructure, and board model in the application layer. Changes were also made accordingly in the rest.module.ts and prisma.module.ts to make sure the new CreateBoard module and repository are properly imported.